requirements.txt
scipy
numpy
jupyter
ipython
easydict
opencv-contrib-python
Cython
pillow
h5py
lmdb
mahotas
pandas
requests
web.py==0.40.dev0
pika
matplotlib
lxml
bs4
cv2